% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/znorm.R
\name{iznorm}
\alias{iznorm}
\title{Inverse Box-Cox Transformation: Calculate Laboratory Values from znorm Values}
\usage{
iznorm(x, limits, lambda = 0, probs = c(0.025, 0.975))
}
\arguments{
\item{x}{\code{numeric}, znorm values.}

\item{limits}{\code{numeric} vector or matrix, lower and upper reference limits.}

\item{lambda}{\code{numeric}, Box-Cox parameter, see \code{znorm}.}

\item{probs}{\code{numeric}, see \code{znorm}.}
}
\value{
\code{numeric}, laboratory values.
}
\description{
Inverse function of \code{znorm}, analogous to \code{iz}/\code{izlog}.
}
\examples{
x <- znorm(1:10, limits = c(2, 8), lambda = 0.5)
}
