% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/znorm.R
\name{znorm}
\alias{znorm}
\title{Normalizing data vector}
\usage{
znorm(x, removeNA = TRUE, ...)
}
\arguments{
\item{x}{Numeric vector.}

\item{removeNA}{A logical value indicating whether NA values should be removed, with default value of TRUE.}

\item{...}{}
}
\value{
A vector containing the normalized input data vector
}
\description{
Normalizing data vector
}
\examples{
znorm(c(16, 9, 13, 5))
}
